feat 数量改为int

+++ src/main/java/com/olivia/peanut/aps/api/entity/apsGoodsForecastUserGoodsData/ApsGoodsForecastUserGoodsDataDto.java
@@ -4,6 +4,7 @@ import com.olivia.peanut.portal.api.entity.BaseEntityDto;
 import lombok.Getter;
 import lombok.Setter;
 
+
 /**
  * (ApsGoodsForecastUserGoodsData)查询对象返回
  *
@@ -18,18 +19,18 @@ public class ApsGoodsForecastUserGoodsDataDto extends BaseEntityDto {
 
   private Long forecastId;
   private Integer year;
-  private Double month_01;
-  private Double month_02;
-  private Double month_03;
-  private Double month_04;
-  private Double month_05;
-  private Double month_06;
-  private Double month_07;
-  private Double month_08;
-  private Double month_09;
-  private Double month_10;
-  private Double month_11;
-  private Double month_12;
+  private Integer month_01;
+  private Integer month_02;
+  private Integer month_03;
+  private Integer month_04;
+  private Integer month_05;
+  private Integer month_06;
+  private Integer month_07;
+  private Integer month_08;
+  private Integer month_09;
+  private Integer month_10;
+  private Integer month_11;
+  private Integer month_12;
 }
